tests

+++ HSEBank/Presentation/FinancialFacade.cs
@@ -38,6 +38,7 @@ public class FinancialFacade : IFinancialFacade
     
     public Operation CreateOperation(OperationDto operationDto)
     {
+        // сделать ли проверку, что есть уже такая операция ?
         if (!_accountFacade.AccountExists(operationDto.BankAccountId))
         {
             throw new ArgumentException($"Account does not exist {nameof(operationDto.BankAccountId)}");
@@ -52,41 +53,73 @@ public class FinancialFacade : IFinancialFacade
 
     public bool EditOperation(EditOperationDto editOperationDto)
     {
+        if (!_categoryFacade.CategoryExists(editOperationDto.CategoryId))
+        {
+            throw new ArgumentException($"Category does not exist {nameof(editOperationDto.CategoryId)}");
+        }
         return _operationFacade.EditOperation(editOperationDto);
     }
 
     public bool DeleteOperation(Guid operationId)
     {
+        if (!_operationFacade.OperationExists(operationId))
+        {
+            throw new ArgumentException($"Operation does not exist {nameof(operationId)}");
+        }
         return _operationFacade.DeleteOperation(operationId);
     }
 
     public Operation GetOperation(Guid operationId)
     {
+        if (!_operationFacade.OperationExists(operationId))
+        {
+            throw new ArgumentException($"Operation does not exist {nameof(operationId)}");
+        }
         return _operationFacade.GetById(operationId);
     }
 
     public BankAccount CreateBankAccount(BankAccountDto bankAccountDto)
     {
+        if (_accountFacade.AccountExists(bankAccountDto.AccountId))
+        {
+            throw new ArgumentException($"Account is already exist {nameof(bankAccountDto.AccountId)}");
+        }
         return _accountFacade.Create(bankAccountDto);
     }
 
     public bool EditBankAccount(EditBankAccountDto editBankAccountDto)
     {
+        if (!_accountFacade.AccountExists(editBankAccountDto.BankAccountId))
+        {
+            throw new ArgumentException($"Account does not exist {nameof(editBankAccountDto.BankAccountId)}");
+        }
         return _accountFacade.EditBankAccount(editBankAccountDto);
     }
 
     public bool DeleteBankAccount(Guid bankAccountId)
     {
+        if (!_accountFacade.AccountExists(bankAccountId))
+        {
+            throw new ArgumentException($"Account does not exist {nameof(bankAccountId)}");
+        }
         return _accountFacade.DeleteBankAccount(bankAccountId);
     }
 
     public BankAccount GetBankAccount(Guid bankAccountId)
     {
+        if (!_accountFacade.AccountExists(bankAccountId))
+        {
+            throw new ArgumentException($"Account does not exist {nameof(bankAccountId)}");
+        }
         return _accountFacade.GetById(bankAccountId);
     }
 
     public Category CreateCategory(CategoryDto categoryDto)
     {
+        if (_categoryFacade.CategoryExists(categoryDto.CategoryId))
+        {
+            throw new ArgumentException($"Category is already exist {nameof(categoryDto.CategoryId)}");
+        }
         return _categoryFacade.Create(categoryDto);
     }
 
@@ -97,16 +130,28 @@ public class FinancialFacade : IFinancialFacade
 
     public bool DeleteCategory(Guid categoryId)
     {
+        if (!_categoryFacade.CategoryExists(categoryId))
+        {
+            throw new ArgumentException($"Category does not exist {nameof(categoryId)}");
+        }
         return _categoryFacade.DeleteCategory(categoryId);
     }
 
     public Category GetCategory(Guid categoryId)
     {
+        if (!_categoryFacade.CategoryExists(categoryId))
+        {
+            throw new ArgumentException($"Category does not exist {nameof(categoryId)}");
+        }
         return _categoryFacade.GetById(categoryId);
     }
 
     public decimal RecalculateBalance(Guid bankAccountId)
     {
+        if (!_accountFacade.AccountExists(bankAccountId))
+        {
+            throw new ArgumentException($"Account does not exist {nameof(bankAccountId)}");
+        }
         decimal balance = 0;
         foreach (var op in _operationFacade.GetByCondition(o => o.BankAccountId == bankAccountId))
         {
